Move Command subclasses into commands directory
I'm trying to get a feel for the code by writing some simple cleanup patches.
