+++ lib/src/commands/stop.dart
@@ -8,8 +8,9 @@ import 'dart:async';
 
 import 'package:args/command_runner.dart';
 import 'package:logging/logging.dart';
-import 'package:sky_tools/src/application_package.dart';
-import 'package:sky_tools/src/device.dart';
+
+import '../application_package.dart';
+import '../device.dart';
 
 final Logger _logging = new Logger('sky_tools.stop');
