+++ test/logs_test.dart
@@ -6,7 +6,7 @@ library logs_test;
 
 import 'package:args/command_runner.dart';
 import 'package:mockito/mockito.dart';
-import 'package:sky_tools/src/logs.dart';
+import 'package:sky_tools/src/commands/logs.dart';
 import 'package:test/test.dart';
 
 import 'src/common.dart';
