+++ test/start_test.dart
@@ -6,7 +6,7 @@ library start_test;
 
 import 'package:args/command_runner.dart';
 import 'package:mockito/mockito.dart';
-import 'package:sky_tools/src/start.dart';
+import 'package:sky_tools/src/commands/start.dart';
 import 'package:test/test.dart';
 
 import 'src/common.dart';
